expand Connector class and clean up ViewController from similarities
